restructure & readme
