+++ src/components/radar/images/stat.png
